Add the unrestricted edition of guidance book
It is said that the contents of the guidance book has been converted to non-copyable pictures, instead of texts since 2023. This unrestricted edition of guidance book created in 2021 enables users to convert it to .doc files or take advantage of the assistance of AI more efficiently. Don't ask me about the password, the password is password.